chore: 모달로직 개선
